versao mais atual sem login

+++ app/Http/Controllers/MovieController.php
@@ -3,9 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\MovieRequest;
-use App\Models\Movie;
-use App\Repository\MovieRepository;
 use App\Service\MovieService;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
@@ -43,7 +42,7 @@ class MovieController extends Controller
             $this->movieService->createMovie($parametrosPermitidos['movie']);
 
             DB::commit();
-            return response()->json(["message" => 'added to your list'],201);
+            return response()->json(["message" => 'added to your list', 'type' => 'success'],201);
         } catch (\Exception $exception){
             DB::rollBack();
             return response($exception->getMessage(),422);
@@ -53,11 +52,17 @@ class MovieController extends Controller
     public function destroy(Request $request)
     {
 
-        $ids = $request->input('ids');
+        $ids = $request->input('id');
+
+        if (empty($ids)) {
+            return response()->json(['message' => 'Please select at least one.', 'type' => 'error']);
+        }
 
         try {
             $this->movieService->delete($ids);
-            return response()->json(['message' => 'Deleted from your list.']);
+            return response()->json(['message' => 'Removed from your list.', 'type' => 'warning']);
+        }catch (ModelNotFoundException $e){
+            return response()->json(['message' => 'movie not found'],404);
         } catch (\Exception $e) {
             return response()->json(['message' => 'Erro ao excluir filmes.', 'error' => $e->getMessage()], 500);
         }
